@@ -3,14 +3,14 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@craco/craco": "7.0.0-alpha.3",
-        "@ethersproject/constants": "^5.6.0",
-        "@thirdweb-dev/react": "^2.0.4",
-        "@thirdweb-dev/sdk": "^2.0.25",
-        "ethers": "^5.6.2",
-        "react": "^18.0.0",
-        "react-dom": "^18.0.0",
-        "react-scripts": "5.0.0"
+        "@craco/craco": "7.0.0",
+        "@ethersproject/constants": "^5.7.0",
+        "@thirdweb-dev/react": "^3.6.9",
+        "@thirdweb-dev/sdk": "^3.6.9",
+        "ethers": "^5.7.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-scripts": "5.0.1"
     },
     "scripts": {
         "start": "craco start",
@@ -37,6 +37,6 @@
         ]
     },
     "devDependencies": {
-        "source-map-loader": "^3.0.1"
+        "source-map-loader": "^4.0.1"
     }
 }
