@@ -1,18 +1,23 @@
 import App from "./App";
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
-import { ChainId, ThirdwebProvider } from "@thirdweb-dev/react";
+// Import thirdweb provider and Goerli ChainId
+import { ChainId } from "@thirdweb-dev/sdk";
+import { ThirdwebProvider } from "@thirdweb-dev/react";
 
 import "./index.css";
 
-const activeChainId = ChainId.Rinkeby;
+// This is the chainId your dApp will work on.
+const activeChainId = ChainId.Goerli;
 
-ReactDOM.render(
-    <React.StrictMode>
-        <ThirdwebProvider desiredChainId={activeChainId}>
-            <App />
-        </ThirdwebProvider>
-    </React.StrictMode>,
-    document.getElementById("root")
+// Wrap your app with the thirdweb provider
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+  <React.StrictMode>
+    <ThirdwebProvider desiredChainId={activeChainId}>
+      <App />
+    </ThirdwebProvider>
+  </React.StrictMode>
 );
