@@ -1,19 +1,21 @@
 import sdk from "./1-initialize-sdk.js";
 import { readFileSync } from "fs";
 
-const editionDrop = sdk.getEditionDrop("0x60d6A1c99BED48d2437Aa291C225dA8d60a37D80");
-
 (async () => {
-    try {
-        await editionDrop.createBatch([
-            {
-                name: "Leaf Village Headband",
-                description: "This NFT will give you access to NarutoDAO!",
-                image: readFileSync("scripts/assets/pokeball.png"),
-            },
-        ]);
-        console.log("✅ Successfully created a new NFT in the drop!");
-    } catch (error) {
-        console.error("failed to create the new NFT", error);
-    }
+  try {
+    const editionDrop = await sdk.getContract(
+      "INSERT_EDITION_DROP_ADDRESS",
+      "edition-drop"
+    );
+    await editionDrop.createBatch([
+      {
+        name: "Leaf Village Headband",
+        description: "This NFT will give you access to NarutoDAO!",
+        image: readFileSync("scripts/assets/headband.png"),
+      },
+    ]);
+    console.log("✅ Successfully created a new NFT in the drop!");
+  } catch (error) {
+    console.error("failed to create the new NFT", error);
+  }
 })();
