@@ -1,31 +1,33 @@
 import sdk from "./1-initialize-sdk.js";
 import { MaxUint256 } from "@ethersproject/constants";
 
-const editionDrop = sdk.getEditionDrop("0x60d6A1c99BED48d2437Aa291C225dA8d60a37D80");
-
 (async () => {
-    try {
-        // We define our claim conditions, this is an array of objects because
-        // we can have multiple phases starting at different times if we want to
-        const claimConditions = [
-            {
-                // When people are gonna be able to start claiming the NFTs (now)
-                startTime: new Date(),
-                // The maximum number of NFTs that can be claimed.
-                maxQuantity: 50_000,
-                // The price of our NFT (free)
-                price: 0,
-                // The amount of NFTs people can claim in one transaction.
-                quantityLimitPerTransaction: 1,
-                // We set the wait between transactions to MaxUint256, which means
-                // people are only allowed to claim once.
-                waitInSeconds: MaxUint256,
-            },
-        ];
+  try {
+    const editionDrop = await sdk.getContract(
+      "INSERT_EDITION_DROP_ADDRESS",
+      "edition-drop"
+    );
+    // We define our claim conditions, this is an array of objects because
+    // we can have multiple phases starting at different times if we want to
+    const claimConditions = [
+      {
+        // When people are gonna be able to start claiming the NFTs (now)
+        startTime: new Date(),
+        // The maximum number of NFTs that can be claimed.
+        maxClaimable: 50_000,
+        // The price of our NFT (free)
+        price: 0,
+        // The amount of NFTs people can claim in one transaction.
+        maxClaimablePerWallet: 1,
+        // We set the wait between transactions to unlimited, which means
+        // people are only allowed to claim once.
+        waitInSeconds: MaxUint256,
+      },
+    ];
 
-        await editionDrop.claimConditions.set("0", claimConditions);
-        console.log("✅ Sucessfully set claim condition!");
-    } catch (error) {
-        console.error("Failed to set claim condition", error);
-    }
+    await editionDrop.claimConditions.set("0", claimConditions);
+    console.log("✅ Sucessfully set claim condition!");
+  } catch (error) {
+    console.error("Failed to set claim condition", error);
+  }
 })();
