@@ -1,19 +1,22 @@
 import sdk from "./1-initialize-sdk.js";
 
-// This is the address of our ERC-20 contract printed out in the step before.
-const token = sdk.getToken("0x29504A69a2A81F9030186de7F2F41Ef4cade1f4a");
-
 (async () => {
-    try {
-        // What's the max supply you want to set? 1,000,000 is a nice number!
-        const amount = 1_000_000;
-        // Interact with your deployed ERC-20 contract and mint the tokens!
-        await token.mintToSelf(amount);
-        const totalSupply = await token.totalSupply();
+  try {
+    // This is the address of our ERC-20 contract printed out in the step before.
+    const token = await sdk.getContract("INSERT_TOKEN_ADDRESS", "token");
+    // What's the max supply you want to set? 1,000,000 is a nice number!
+    const amount = 1_000_000;
+    // Interact with your deployed ERC-20 contract and mint the tokens!
+    await token.mint(amount);
+    const totalSupply = await token.totalSupply();
 
-        // Print out how many of our token's are out there now!
-        console.log("✅ There now is", totalSupply.displayValue, "$HOKAGE in circulation");
-    } catch (error) {
-        console.error("Failed to print money", error);
-    }
+    // Print out how many of our token's are out there now!
+    console.log(
+      "✅ There now is",
+      totalSupply.displayValue,
+      "$HOKAGE in circulation"
+    );
+  } catch (error) {
+    console.error("Failed to print money", error);
+  }
 })();
